update: routes/web

+++ routes/web.php
@@ -4,9 +4,9 @@ use App\Http\Controllers\AlternatifController;
 use App\Http\Controllers\DashboardController;
 use App\Http\Controllers\KriteriaController;
 use App\Http\Controllers\LoginController;
-use App\Http\Controllers\PenilaianController;
-use App\Http\Controllers\PerhitunganController;
+use App\Http\Controllers\PerhitunganAlternatifController;
 use App\Http\Controllers\PerhitunganKriteriaController;
+use App\Http\Controllers\RankingController;
 use App\Http\Controllers\RegisterController;
 use Illuminate\Support\Facades\Route;
 
@@ -39,30 +39,30 @@ Route::controller(DashboardController::class)->name('dashboard.')->middleware('a
     Route::post('/dashboard', 'store')->name('store');
 });
 
-Route::controller(KriteriaController::class)->name('kriteria.')->middleware('auth', 'user-role:manajer,pimpinan')->group(function () {
-    Route::get('/dashboard/data-kriteria', 'index')->name('index');
-    Route::get('/dashboard/data-kriteria/view-kriteria/{id}', 'show')->name('show');
+Route::controller(KriteriaController::class)->name('kriteria.')->middleware('auth')->group(function () {
+    Route::get('/dashboard/data-kriteria', 'index')->name('index')->middleware('user-role:manajer,pimpinan');
+    Route::get('/dashboard/data-kriteria/view-kriteria/{id}', 'show')->name('show')->middleware('user-role:manajer,pimpinan');
 
-    Route::get('/dashboard/data-kriteria/tambah-kriteria', 'create')->name('create');
-    Route::post('/dashboard/data-kriteria', 'store')->name('store');
+    Route::get('/dashboard/data-kriteria/tambah-kriteria', 'create')->name('create')->middleware('user-role:manajer');
+    Route::post('/dashboard/data-kriteria', 'store')->name('store')->middleware('user-role:manajer');
 
-    Route::get('/dashboard/data-kriteria/ubah-kriteria/{id}/edit', 'edit')->name('edit');
-    Route::put('/dashboard/data-kriteria/{id}', 'update')->name('update');
+    Route::get('/dashboard/data-kriteria/ubah-kriteria/{id}/edit', 'edit')->name('edit')->middleware('user-role:manajer');
+    Route::put('/dashboard/data-kriteria/{id}', 'update')->name('update')->middleware('user-role:manajer');
 
-    Route::delete('/dashboard/data-kriteria/{id}', 'destroy')->name('destroy');
+    Route::delete('/dashboard/data-kriteria/{id}', 'destroy')->name('destroy')->middleware('user-role:manajer');
 });
 
 Route::controller(AlternatifController::class)->name('alternatif.')->middleware('auth', 'user-role:manajer,pimpinan')->group(function () {
-    Route::get('/dashboard/data-alternatif', 'index')->name('index');
-    Route::get('/dashboard/data-alternatif/view-alternatif/{id}', 'show')->name('show');
+    Route::get('/dashboard/data-alternatif', 'index')->name('index')->middleware('user-role:manajer,pimpinan');
+    Route::get('/dashboard/data-alternatif/view-alternatif/{id}', 'show')->name('show')->middleware('user-role:manajer,pimpinan');
 
-    Route::get('/dashboard/data-alternatif/tambah-alternatif', 'create')->name('create');
-    Route::post('/dashboard/data-alternatif', 'store')->name('store');
+    Route::get('/dashboard/data-alternatif/tambah-alternatif', 'create')->name('create')->middleware('user-role:manajer');
+    Route::post('/dashboard/data-alternatif', 'store')->name('store')->middleware('user-role:manajer');
 
-    Route::get('/dashboard/data-alternatif/ubah-alternatif/{id}/edit', 'edit')->name('edit');
-    Route::put('/dashboard/data-alternatif/{id}', 'update')->name('update');
+    Route::get('/dashboard/data-alternatif/ubah-alternatif/{id}/edit', 'edit')->name('edit')->middleware('user-role:manajer');
+    Route::put('/dashboard/data-alternatif/{id}', 'update')->name('update')->middleware('user-role:manajer');
 
-    Route::delete('/dashboard/data-alternatif/{id}', 'destroy')->name('destroy');
+    Route::delete('/dashboard/data-alternatif/{id}', 'destroy')->name('destroy')->middleware('user-role:manajer');
 });
 
 Route::controller(PerhitunganKriteriaController::class)->name('perhitunganKriteria.')->middleware('auth', 'user-role:manajer')->group(function () {
@@ -70,4 +70,16 @@ Route::controller(PerhitunganKriteriaController::class)->name('perhitunganKriter
     Route::get('/dashboard/perhitungan-perbandingan-kriteria/hasil-perbandingan-kriteria', 'hasil')->name('hasil');
 
     Route::post('/dashboard/perhitungan-perbandingan-kriteria', 'store')->name('store');
+});
+
+Route::controller(PerhitunganAlternatifController::class)->name('perhitunganAlternatif.')->middleware('auth', 'user-role:manajer')->group(function () {
+    Route::get('/dashboard/perhitungan-perbandingan-alternatif', 'index')->name('index');
+    Route::get('/dashboard/perhitungan-perbandingan-alternatif/hasil-perbandingan-alternatif', 'hasil')->name('hasil');
+
+    Route::post('/dashboard/perhitungan-perbandingan-alternatif', 'store')->name('store');
+});
+
+Route::controller(RankingController::class)->name('ranking.')->middleware('auth', 'user-role:manajer,pimpinan')->group(function () {
+    Route::get('/dashboard/perankingan', 'index')->name('index');
+    Route::post('/dashboard/perankingan', 'store')->name('store');
 });
